requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.1
django-allauth==0.34.0
gunicorn==19.7.1
idna==2.6
kombu==4.1.0
oauthlib==2.0.6
psycopg2==2.7.3.2
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
vine==1.1.4
